PR comment resolve
