+++ nautobot/core/api/urls.py
@@ -37,7 +37,7 @@ ui_api_patterns = [
     path("core/", include((core_api_patterns, "core-api"))),
     path("get-menu/", GetMenuAPIView.as_view(), name="get-menu"),
     path("get-object-counts/", GetObjectCountsView.as_view(), name="get-object-counts"),
-    path("get-settings/", GetSettingsView.as_view(), name="get-settings"),
+    path("settings/", GetSettingsView.as_view(), name="settings"),
 ]
 
 urlpatterns = [
