+++ nautobot/core/api/views.py
@@ -883,7 +883,6 @@ class GetSettingsView(NautobotAPIVersionMixin, APIView):
 
     @extend_schema(exclude=True)
     def get(self, request):
-
         # As of now, we just have `allowed_settings` settings. Because the purpose of this API is limited for the time being,
         # we may need more access to serializable and non-serializable settings data as the new UI expands. As a result,
         # exposing all settings data would be desirable in the future.
@@ -891,17 +890,17 @@ class GetSettingsView(NautobotAPIVersionMixin, APIView):
         # e.g `SECRET_KEY`, `NAPALM_PASSWORD` e.t.c. also `allowed_settings` can be moved into settings.py
         allowed_settings = ["FEEDBACK_BUTTON_ENABLED"]
         # Filter out settings_names not allowed to be exposed to the API
-        filter_params = [name for name in request.GET.getlist("name") if name in allowed_settings]
-        serializable_settings = {}
+        valid_settings = [name for name in request.GET.getlist("name") if name in allowed_settings]
 
-        if filter_params:
-            for settings_name in filter_params:
-                # Not using `getattr(settings, settings_name, None)` directly as `settings_name` might exists in `settings` with its value as None
-                if hasattr(settings, settings_name):
-                    value = getattr(settings, settings_name)
-                    if self.is_serializable(value):
-                        serializable_settings[settings_name] = value
-        return Response(serializable_settings)
+        invalid_settings = [name for name in request.GET.getlist("name") if name not in allowed_settings]
+        if invalid_settings:
+            return Response(
+                {"error": f"Invalid settings names specified: {', '.join(invalid_settings)}."},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+        settings = {settings_name: get_settings_or_config(settings_name) for settings_name in valid_settings}
+        return Response(settings)
 
 
 #
