+++ nautobot/ui/package.json
@@ -1,6 +1,6 @@
 {
     "name": "nautobot",
-    "version": "2.0.0-alpha.1",
+    "version": "2.0.0-beta.2",
     "private": true,
     "proxy": "http://nautobot:8080",
     "dependencies": {
